meta-tegra: fix rcu-stalls when kernel enable CONFIG_SHADOW_CALL_STACK
Issue: LIN1025-2680

On aarch64, the instrumentation makes use of the x18 register to reference
the shadow call stack, meaning that references to the shadow call stack don't
have to be stored in memory.

This feature on aarch64 also need to be supported by compiler, for gcc, it
need to turn on the option -ffixed-x18 [1]

Since NV modset and display driver is out of tree kernel module, so we need to
set gcc option in their Makefile.

Ref:
[1] https://gcc.gnu.org/pipermail/gcc-patches/2022-January/588946.html

(LOCAL REV: NOT UPSTREAM) - WRLinux specific

Signed-off-by: Xiangyu Chen <xiangyu.chen@windriver.com>